# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /legs/{id}/dtmf:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: sendDTMF
      summary: Send DTMF digits on a leg
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DTMFRequest'
      responses:
        '200':
          description: Digits sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: sent
        '400':
          description: Invalid JSON or empty digits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '500':
          description: DTMF writer unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /legs/{id}/agent:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: agentLeg
      summary: Attach an AI agent to a leg
      description: >
        Bridges audio bidirectionally with an ElevenLabs conversational AI agent.
        Standalone legs use direct audio; legs in a room use mixer taps.
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '200':
          description: Agent started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: agent_started
                  leg_id:
                    type: string
                required: [status, leg_id]
        '400':
          description: Invalid JSON or missing agent_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg not connected, agent already attached, or no audio reader/writer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No ElevenLabs API key provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopAgentLeg
      summary: Detach the agent from a leg
      tags: [Legs]
      responses:
        '200':
          description: Agent stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: agent_stopped
        '404':
          description: No agent attached to this leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Rooms ───────────────────────────────────────────────────────────────
...
  /rooms/{id}/agent:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: agentRoom
      summary: Attach an AI agent to a room
      description: >
        The agent joins as a virtual participant, hearing all participants
        (mixed-minus-self) and speaking to everyone.
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '200':
          description: Agent started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: agent_started
                  room_id:
                    type: string
                required: [status, room_id]
        '400':
          description: Invalid JSON or missing agent_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: Agent already attached to this room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No ElevenLabs API key provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopAgentRoom
      summary: Detach the agent from a room
      tags: [Rooms]
      responses:
        '200':
          description: Agent stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: agent_stopped
        '404':
          description: No agent attached to this room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    LegId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Leg ID
    RoomId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Room ID
...
  responses:
    LegNotFound:
      description: Leg not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ── Request bodies ──────────────────────────────────────────────────
...
    DTMFRequest:
      type: object
      properties:
        digits:
          type: string
          description: 'DTMF digits to send (0-9, *, #)'
      required: [digits]
...
    AgentRequest:
      type: object
      properties:
        agent_id:
          type: string
          description: ElevenLabs agent ID
        first_message:
          type: string
          description: Override the agent's first message
        language:
          type: string
          description: Language code
        dynamic_variables:
          type: object
          additionalProperties:
            type: string
          description: Key-value pairs passed to the agent
        api_key:
          type: string
          description: ElevenLabs API key override
      required: [agent_id]
...